feat: overwrite files

+++ orch.core.ef/System/EFSystemDatabase.cs
@@ -42,7 +42,18 @@ namespace orch.core.ef.System
 
         public void CreateFile(ContentFile cf)
         {
+            var dalFile = new DALContentFile(cf);
             _dbContext.Files.Add(new DALContentFile(cf));
+            _dbContext.SaveChanges();
+            _dbContext.Entry(dalFile).State = EntityState.Detached;
+        }
+
+        public void DeleteFile(Guid fileId)
+        {
+            if (!_dbContext.Files.Any(f => f.FileId == fileId))
+                throw new FileNotFoundException($"File {fileId} doesn't exist");
+
+            _dbContext.Files.Remove(_dbContext.Files.First(f => f.FileId == fileId));
             _dbContext.SaveChanges();
         }
 
@@ -75,9 +86,11 @@ namespace orch.core.ef.System
 
         public ContentFile? GetFile(Guid file_id)
         {
-            return _dbContext.Files.Where(file => file.FileId == file_id)
-                          .Select(file => new ContentFile(file))
-                          .FirstOrDefault();
+            return _dbContext.Files
+                .AsNoTracking()
+                .Where(file => file.FileId == file_id)
+                .Select(file => new ContentFile(file))
+                .FirstOrDefault();
         }
 
         public PagedList<ContentFile> GetFiles(int pageNumber, int pageSize, ContentFileFilter? filter = null)
@@ -133,7 +146,7 @@ namespace orch.core.ef.System
                 .ToList();
         }
 
-        public ContentFile SaveFile(string fileName, Stream r, Guid? fileId = null)
+        public ContentFile SaveFile(string fileName, Stream r, Guid? fileId = null, bool overwrite = false)
         {
             if (r.Length > _contentServerConfig.MaxFileSize)
                 throw new ArgumentException($"File size ({r.Length / (1024.0 * 1024.0):F2} MB) exceeds the maximum allowed size ({_contentServerConfig.MaxFileSize / (1024.0 * 1024.0):F2} MB)");
@@ -145,6 +158,8 @@ namespace orch.core.ef.System
                 MimeType = MimeMapping.MimeUtility.GetMimeMapping(fileName), //TODO: check if we can use f.ContentType
                 CreateTime = _host.CurrentTime()
             };
+
+            var existingFile = GetFile(item.FileId);
 
             // Create directory if it does not already exist
             if (!Directory.Exists(_contentServerConfig.BaseDir))
@@ -160,6 +175,9 @@ namespace orch.core.ef.System
 
             // Create file path for saving the file
             var filePath = Path.Combine(_contentServerConfig.BaseDir, $"{item.FileId}.content");
+
+            if (File.Exists(filePath) && !overwrite)
+                throw new InvalidOperationException($"File '{item.FileId}' already exists on file system");
 
             // Create file and write stream data to it
             using (var os = File.Create(filePath))
@@ -181,18 +199,38 @@ namespace orch.core.ef.System
             hashFunc.TransformFinalBlock(buffer, 0, 0);
             hashFunc.Dispose();
 
+            var transaction = _dbContext.Database.BeginTransaction();
+
             try
             {
-                // Create entry for file in the database and return ContentFile object representing the saved file
+                if (existingFile is not null)
+                {
+                    if (overwrite)
+                    {
+                        DeleteFile(existingFile.FileId);
+                    }
+                    else
+                    {
+                        throw new InvalidOperationException(
+                            $"Content index for file '{existingFile.FileId}' already exists");
+                    }
+                }
+
                 CreateFile(item);
+
+                transaction.Commit();
+                
                 return item;
             }
             catch
             {
+                transaction.Rollback();
+                
                 // If creating entry in database fails, delete the saved file and throw exception
                 try
                 {
-                    File.Delete(filePath);
+                    if (existingFile is not null)
+                        File.Delete(filePath);
                 }
                 catch (Exception deleteException)
                 {
